% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_pow.R
\name{my_pow}
\alias{my_pow}
\title{power function}
\usage{
my_pow(x, power = 2)
}
\arguments{
\item{x}{input to be converted to the \code{power}}
}
\value{
Numeric representing \code{x} raised to the power of \code{power}
}
\description{
This function power the input.
}
\examples{
my_power(4)
my_power(4, power =2)

}
\keyword{temperature}
