% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_rf_cv.R
\name{my_rf_cv}
\alias{my_rf_cv}
\title{Random Forest Cross-Validation}
\usage{
my_rf_cv(k)
}
\arguments{
\item{k}{number of folds}
}
\value{
a numeric with the cross-validation error
}
\description{
This function predict output using Random Forest Cross-Validation algorithms.
}
\examples{
my_result_1 <- my_knn_cv(train = penguins, cl = penguins$species, k_nn = 1, k_cv = 5 )
my_result_2 <- my_knn_cv(train = penguins, cl = penguins$species, k_nn = 5, k_cv = 5 )
train_err_1 <- sum(as.numeric(penguins$species != my_result_1$class)) / nrow(penguins)
train_err_2 <- sum(as.numeric(penguins$species != my_result_2$class)) / nrow(penguins)
my_table_row_1 <- cbind("k_nn = 1" = my_result_1$cv_err, "k_nn = 5" = my_result_2$cv_err)
my_table_row_2 <- cbind("k_nn = 1" = train_err_1, "k_nn = 5" = train_err_2)
my_table <- rbind(my_table_row_1, my_table_row_2)
rownames(my_table) <- c("cv_err", "training_err")
my_table

}
\keyword{{prediction}}
